+++ app/src/main/java/com/dimowner/goodweather/data/remote/model/Wind.kt
@@ -27,7 +27,7 @@ import com.google.gson.annotations.SerializedName
  */
 data class Wind(
 		@SerializedName("speed")
-		val speed: Int,
+		val speed: Float,
 		@SerializedName("deg")
-		val deg: Int
+		val deg: Float
 )
